flask_deploy/requirements.txt
keras
tensorflow
flask
pillow